+++ CHANGELOG.md
@@ -35,6 +35,7 @@ at anytime.
   * unittests for `ComponentManager`
   * script to generate docs/api.json file (https://github.com/lbryio/lbry.tech/issues/42)
   * additional information to the balance error message when editing a claim (https://github.com/lbryio/lbry/pull/1309)
+  * `address` and `port` arguments to `peer_ping` (https://github.com/lbryio/lbry/issues/1313)
 
 ### Removed
   * most of the internal attributes from `Daemon`